cwfs: fix 1GB memsize limitation
the malloc pool allocator is limited in its allocation
size. as almost all data structures in cwfs are never
freed, use brk() in ialloc() instead of mallocalign().
this means memory returned by ialloc() cannot be freed!

to make sure we do not call free by accident, remove
the #define malloc(n) ialloc(n, 0) macro and use ialloc()
directly as in the original code to show the intend
of permanent allocations.

+++ sys/src/cmd/cwfs/iobuf.c
@@ -1,7 +1,7 @@
 #include	"all.h"
 #include	"io.h"
 
-extern	long nhiob;
+extern	uint nhiob;
 extern	Hiob *hiob;
 
 Iobuf*
